refactor(6648): split governance extraction modules
