+++ crates/kamn-core/tests/kamn_core_governance_phase1_contract.rs
@@ -6,6 +6,50 @@ fn repo_file(path: &str) -> String {
     fs::read_to_string(full).unwrap_or_default()
 }
 
+fn assert_governance_lib_declares_modules(governance_lib: &str) {
+    for marker in [
+        "pub mod governance_workflow;",
+        "pub mod operator_binding;",
+        "pub mod operator_actions;",
+    ] {
+        assert!(
+            governance_lib.contains(marker),
+            "kamn-governance lib should declare extracted module: {marker}"
+        );
+    }
+}
+
+fn assert_core_shims_reexport_governance_crate() {
+    let workflow_shim = repo_file("crates/kamn-core/src/governance_workflow.rs");
+    let binding_shim = repo_file("crates/kamn-core/src/operator_binding.rs");
+    let actions_shim = repo_file("crates/kamn-core/src/operator_actions.rs");
+    assert!(
+        workflow_shim.contains("pub use kamn_governance::governance_workflow::*;"),
+        "kamn-core governance_workflow shim should re-export from kamn_governance"
+    );
+    assert!(
+        binding_shim.contains("pub use kamn_governance::operator_binding::*;"),
+        "kamn-core operator_binding shim should re-export from kamn_governance"
+    );
+    assert!(
+        actions_shim.contains("pub use kamn_governance::operator_actions::*;"),
+        "kamn-core operator_actions shim should re-export from kamn_governance"
+    );
+}
+
+fn assert_dashboards_remain_in_core() {
+    let dashboard_api = repo_file("crates/kamn-core/src/operator_dashboard_api.rs");
+    let dashboard_ui = repo_file("crates/kamn-core/src/operator_dashboard_ui.rs");
+    assert!(
+        dashboard_api.contains("pub struct OperatorDashboardApi"),
+        "operator_dashboard_api should remain implemented in kamn-core during phase1"
+    );
+    assert!(
+        dashboard_ui.contains("pub struct OperatorDashboardUi"),
+        "operator_dashboard_ui should remain implemented in kamn-core during phase1"
+    );
+}
+
 #[test]
 fn governance_phase1_contract_declares_workspace_and_core_dependencies() {
     let workspace = repo_file("Cargo.toml");
@@ -29,44 +73,9 @@ fn governance_phase1_contract_declares_workspace_and_core_dependencies() {
 #[test]
 fn governance_phase1_contract_extracts_governance_primitives_and_keeps_dashboards_in_core() {
     let governance_lib = repo_file("crates/kamn-governance/src/lib.rs");
-    let workflow_shim = repo_file("crates/kamn-core/src/governance_workflow.rs");
-    let binding_shim = repo_file("crates/kamn-core/src/operator_binding.rs");
-    let actions_shim = repo_file("crates/kamn-core/src/operator_actions.rs");
-    let dashboard_api = repo_file("crates/kamn-core/src/operator_dashboard_api.rs");
-    let dashboard_ui = repo_file("crates/kamn-core/src/operator_dashboard_ui.rs");
-
-    for marker in [
-        "pub mod governance_workflow;",
-        "pub mod operator_binding;",
-        "pub mod operator_actions;",
-    ] {
-        assert!(
-            governance_lib.contains(marker),
-            "kamn-governance lib should declare extracted module: {marker}"
-        );
-    }
-
-    assert!(
-        workflow_shim.contains("pub use kamn_governance::governance_workflow::*;"),
-        "kamn-core governance_workflow shim should re-export from kamn_governance"
-    );
-    assert!(
-        binding_shim.contains("pub use kamn_governance::operator_binding::*;"),
-        "kamn-core operator_binding shim should re-export from kamn_governance"
-    );
-    assert!(
-        actions_shim.contains("pub use kamn_governance::operator_actions::*;"),
-        "kamn-core operator_actions shim should re-export from kamn_governance"
-    );
-
-    assert!(
-        dashboard_api.contains("pub struct OperatorDashboardApi"),
-        "operator_dashboard_api should remain implemented in kamn-core during phase1"
-    );
-    assert!(
-        dashboard_ui.contains("pub struct OperatorDashboardUi"),
-        "operator_dashboard_ui should remain implemented in kamn-core during phase1"
-    );
+    assert_governance_lib_declares_modules(&governance_lib);
+    assert_core_shims_reexport_governance_crate();
+    assert_dashboards_remain_in_core();
 }
 
 #[test]
