+++ crates/kamn-governance/src/operator_binding/constants.rs
@@ -1,0 +1,7 @@
+pub(super) const HUMAN_DID_PREFIX: &str = "kamn:did:human:";
+pub(super) const AGENT_DID_PREFIX: &str = "kamn:did:agent:";
+pub(super) const CANONICAL_PROOF_TYPE: &str = "Ed25519Signature2020";
+pub(super) const OPERATOR_BINDING_INVALID_AGENT_DID_REASON_CODE: &str =
+    "operator_binding_invalid_agent_did";
+pub(super) const OPERATOR_BINDING_INVALID_OPERATOR_DID_REASON_CODE: &str =
+    "operator_binding_invalid_operator_did";
